+++ test/__scenarios__/z-record.partial-type.input.ts
@@ -1,0 +1,4 @@
+import { z } from "zod";
+
+const MyRecord = z.record(z.enum(["a", "b", "c"]), z.number());
+type MyRecord = z.infer<typeof MyRecord>;